[Feat] other: 네비게이팅
