+++ QueenCam/QueenCam/Presentation/Views/Common/Navigation/NavigationRouter.swift
@@ -1,0 +1,27 @@
+//
+//  NavigationRouter.swift
+//  QueenCam
+//
+//  Created by 임영택 on 2/14/26.
+//
+
+import SwiftUI
+
+@Observable
+final class NavigationRouter {
+  var path: [Route] = []
+
+  func push(_ route: Route) {
+    path.append(route)
+  }
+
+  func pop() {
+    if !path.isEmpty {
+      path.removeLast()
+    }
+  }
+
+  func reset() {
+    path = []
+  }
+}
